Fix: App crashing in ts template due to change, and move helpers to lib for all templates
