+++ js-cjs-then/src/server.js
@@ -1,6 +1,6 @@
 const mongooseConnect = require('./db.connect');
 const { app } = require('./app');
-const { connectionType } = require('../node-mongo-helpers');
+const { connectionType } = require('./lib/helpers');
 
 const port = connectionType().port || 3000;
