+++ js-esm-async-await/src/db.connect.js
@@ -1,5 +1,5 @@
 import mongoose from 'mongoose';
-import { connectionType, afterDBconnectSuccessful, connectToDBunsuccessful } from '../node-mongo-helpers';
+import { connectionType, afterDBconnectSuccessful, connectToDBunsuccessful } from './lib/helpers';
 
 const mongooseConnect = async (port) => {
   try {
