+++ js-esm-async-await/src/lib/helpers/index.js
@@ -1,5 +1,5 @@
 import { spawn } from 'child_process';
-import package_json from './package.json';
+import package_json from '../../../package.json';
 import chalk from 'chalk';
 import dotenv from 'dotenv';
 
@@ -8,15 +8,15 @@ dotenv.config();
 /* eslint-disable no-console */
 
 // console
-export const success = (message: string) => {
+export const success = (message) => {
   console.log( chalk.greenBright(message) );
 }
 
-export const warning = (message: string) => {
+export const warning = (message) => {
   console.log( chalk.yellowBright(message) );
 }
 
-export const error = (message: string) => {
+export const error = (message) => {
   console.log( chalk.redBright(message) );
 }
 
@@ -29,50 +29,47 @@ const startScript = {
 };
 
 export const connectionType = () => {
-  let connectionChoice: {
-    port: string | number;
-    uri: string;
-  } = { port: '', uri: '' };
+  let connectionChoice = { port: '', uri: '' };
 
   if (startScript.atlas) {
     connectionChoice = {
-      port: process.env.PORT_ATLAS as string | number,
-      uri: process.env.MONGODB_ATLAS_URI as string,
+      port: process.env.PORT_ATLAS,
+      uri: process.env.MONGODB_ATLAS_URI,
     };
   }
 
   if (startScript.local) {
     connectionChoice = {
-      port: process.env.PORT_LOCAL as string | number,
-      uri: process.env.MONGODB_LOCAL_URI as string,
+      port: process.env.PORT_LOCAL,
+      uri: process.env.MONGODB_LOCAL_URI,
     };
   }
 
   return connectionChoice;
 }
 
-
 // DB connect
-export const npmRunPackageJsonScript = ({ script, currentWorkingDir } : { script: string, currentWorkingDir: string }): void => {
+export const watchEslint = () => {
   const npm = process.platform === 'win32' ? 'npm.cmd' : 'npm';
-  spawn(npm, ['run', script], { cwd: currentWorkingDir, stdio: 'inherit' });
+  spawn(npm, ['run', 'lint:watch'], { cwd: './', stdio: 'inherit' });
 }
 
-export const server = (serverPort: number | string): void => {
+export const server = (serverPort) => {
   try {
-    success(`\nnode-mongo (Typescript) API boilerplate template v${package_json.version}`);
+    const description = package_json.description.replace(' generated via Collabo Community\'s create-collabo-app project', '');
+    success(`\nv${package_json.version} ${description}\n\nGenerated via Collabo Community's create-collabo-app project`);
     success(`\nServer running at ${serverPort}`);
   } catch (err) {
     error(`${{ err }}`);
   }
 }
 
-const eslintAndServer = (serverPort: number | string) => {
-  npmRunPackageJsonScript({ script: 'lint:watch', currentWorkingDir: './' });
+const eslintAndServer = (serverPort) => {
+  watchEslint();
   server(serverPort);
 }
 
-export const afterDBconnectSuccessful = (serverPort: number | string) => {
+export const afterDBconnectSuccessful = (serverPort) => {
   const serverType = {
     atlas: startScript.atlas ? 'ATLAS' : '',
     local: startScript.local ? 'LOCAL ' : '',
@@ -81,7 +78,7 @@ export const afterDBconnectSuccessful = (serverPort: number | string) => {
   eslintAndServer(serverPort);
 }
 
-export const connectToDBunsuccessful = (err: { message: unknown; }) => {
+export const connectToDBunsuccessful = (err) => {
   error(`\nError in DB connection: ${err.message}\n`);
   warning('Refer to the node-mongo documentation: https://code-collabo.gitbook.io/node-mongo-v2.0.0\n\nGet further help from Code Collabo Community\'s Node mongo channel: https://matrix.to/#/#collabo-node-mongo:gitter.im\n');
 }
