+++ js-esm-async-await/src/server.js
@@ -1,6 +1,6 @@
 import mongooseConnect from './db.connect';
 import { app as app } from './app';
-import { connectionType } from '../node-mongo-helpers';
+import { connectionType } from './lib/helpers';
 
 const port = connectionType().port || 3000;
