+++ ts-esm-async-await/src/api/controllers/app.controller.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express';
-import { success } from '../../node-mongo-helpers';
+import { success } from '../../lib/helpers';
 
 export const getAppController = async (req: Request, res: Response) => {
   const message = 'App works!';
