+++ ts-esm-async-await/src/lib/errors/ErrorHandler.ts
@@ -1,6 +1,6 @@
 import { Response } from 'express';
 import { CustomErrorInterface, HttpCode } from './CustomError';
-import { error } from '../../node-mongo-helpers';
+import { error } from '../helpers';
 
 class ErrorHandler {
