+++ ts-esm-async-await/src/lib/helpers/index.ts
@@ -1,5 +1,5 @@
 import { spawn } from 'child_process';
-import package_json from './package.json';
+import package_json from '../../../package.json';
 import chalk from 'chalk';
 import dotenv from 'dotenv';
 
@@ -8,15 +8,15 @@ dotenv.config();
 /* eslint-disable no-console */
 
 // console
-export const success = (message) => {
+export const success = (message: string) => {
   console.log( chalk.greenBright(message) );
 }
 
-export const warning = (message) => {
+export const warning = (message: string) => {
   console.log( chalk.yellowBright(message) );
 }
 
-export const error = (message) => {
+export const error = (message: string) => {
   console.log( chalk.redBright(message) );
 }
 
@@ -29,46 +29,51 @@ const startScript = {
 };
 
 export const connectionType = () => {
-  let connectionChoice = { port: '', uri: '' };
+  let connectionChoice: {
+    port: string | number;
+    uri: string;
+  } = { port: '', uri: '' };
 
   if (startScript.atlas) {
     connectionChoice = {
-      port: process.env.PORT_ATLAS,
-      uri: process.env.MONGODB_ATLAS_URI,
+      port: process.env.PORT_ATLAS as string | number,
+      uri: process.env.MONGODB_ATLAS_URI as string,
     };
   }
 
   if (startScript.local) {
     connectionChoice = {
-      port: process.env.PORT_LOCAL,
-      uri: process.env.MONGODB_LOCAL_URI,
+      port: process.env.PORT_LOCAL as string | number,
+      uri: process.env.MONGODB_LOCAL_URI as string,
     };
   }
 
   return connectionChoice;
 }
 
+
 // DB connect
-export const watchEslint = () => {
+export const npmRunPackageJsonScript = ({ script, currentWorkingDir } : { script: string, currentWorkingDir: string }): void => {
   const npm = process.platform === 'win32' ? 'npm.cmd' : 'npm';
-  spawn(npm, ['run', 'lint:watch'], { cwd: './', stdio: 'inherit' });
+  spawn(npm, ['run', script], { cwd: currentWorkingDir, stdio: 'inherit' });
 }
 
-export const server = (serverPort) => {
+export const server = (serverPort: number | string): void => {
   try {
-    success(`\nnode-mongo (ES Module) API boilerplate template v${package_json.version}`);
+    const description: string = package_json.description.replace(' generated via Collabo Community\'s create-collabo-app project', '');
+    success(`\nv${package_json.version} ${description}\n\nGenerated via Collabo Community's create-collabo-app project`);
     success(`\nServer running at ${serverPort}`);
   } catch (err) {
     error(`${{ err }}`);
   }
 }
 
-const eslintAndServer = (serverPort) => {
-  watchEslint();
+const eslintAndServer = (serverPort: number | string) => {
+  npmRunPackageJsonScript({ script: 'lint:watch', currentWorkingDir: './' });
   server(serverPort);
 }
 
-export const afterDBconnectSuccessful = (serverPort) => {
+export const afterDBconnectSuccessful = (serverPort: number | string) => {
   const serverType = {
     atlas: startScript.atlas ? 'ATLAS' : '',
     local: startScript.local ? 'LOCAL ' : '',
@@ -77,7 +82,7 @@ export const afterDBconnectSuccessful = (serverPort) => {
   eslintAndServer(serverPort);
 }
 
-export const connectToDBunsuccessful = (err) => {
+export const connectToDBunsuccessful = (err: { message: unknown; }) => {
   error(`\nError in DB connection: ${err.message}\n`);
   warning('Refer to the node-mongo documentation: https://code-collabo.gitbook.io/node-mongo-v2.0.0\n\nGet further help from Code Collabo Community\'s Node mongo channel: https://matrix.to/#/#collabo-node-mongo:gitter.im\n');
 }
